requirements.txt
appdirs==1.4.4
certifi==2023.11.17
colorama==0.4.6
importlib-metadata==7.0.1
MouseInfo==0.1.3
packaging==23.2
pillow==10.2.0
PyAutoGUI==0.9.54
pyee==8.2.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pyppeteer==1.0.2
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
python-dotenv==1.0.0
pytweening==1.0.7
tqdm==4.66.1
urllib3==1.26.18
websockets==10.4
zipp==3.17.0
